+++ WebApplication/src/components/StudyDetails.vue
@@ -25,7 +25,7 @@ export default {
         this.allLabelsLocalCopy = await api.loadAllLabels();
     },
     async mounted() {
-        this.samePatientStudiesCount = (await api.getSamePatientStudies(this.patientMainDicomTags['PatientID'])).length;
+        this.samePatientStudiesCount = (await api.getSamePatientStudies(this.patientMainDicomTags, this.uiOptions.ShowSamePatientStudiesFilter)).length;
         this.hasLoadedSamePatientsStudiesCount = true;
         Tags.init();
     },
@@ -35,7 +35,21 @@ export default {
         }),
         selectedValues() {
             return this.labelsModel.join(",");
+        },
+        sampePatientStudiesLink() {
+            let filters = [];
+            for (let tag of this.uiOptions.ShowSamePatientStudiesFilter) {
+                if (tag in this.patientMainDicomTags) {
+                    if (tag in ["PatientBirthDate"]) {
+                        filters.push(tag + "=" + this.patientMainDicomTags[tag] + "");
+                    } else {
+                        filters.push(tag + "=\"" + this.patientMainDicomTags[tag] + "\"");
+                    }
+                }
+            }
+            return "/filtered-studies?" + filters.join('&');
         }
+        
     },
     components: { SeriesItem, SeriesList, ResourceButtonGroup, ResourceDetailText },
     methods: {
@@ -101,7 +115,7 @@ export default {
                 </ul>
                 <p v-if="hasLoadedSamePatientsStudiesCount && samePatientStudiesCount > 1">
                     {{ $t('this_patient_has_other_studies', { count: samePatientStudiesCount }) }}.
-                    <router-link v-bind:to='"/filtered-studies?PatientID=\"" + patientMainDicomTags["PatientID"] + "\""' >
+                    <router-link v-bind:to='sampePatientStudiesLink' >
                         {{ $t('this_patient_has_other_studies_show') }}
                     </router-link>
                 </p>