+++ release-notes.md
@@ -5,6 +5,8 @@ Changes:
 - added a button to download multiple studies at once (only available if running an
   Orthanc version > 1.12.1 - not officially released at the time this plugin was
   released)
+- new configuration "UiOptions.ShowSamePatientStudiesFilter" to list the tags that 
+  are used to identify studies belonging to the same patient.
 - added Russian translations